Snuoperasjon. På tide med arbeid for Jon. Kos deg :) opprett environment er eneste som er funky og oppretter configdata i tillegg.... to be fixed...

+++ Domain/Repositories/IConfigdataRepository.cs
@@ -18,7 +18,7 @@ namespace Domain.Repositories
 
         IEnumerable<Environment> Environments { get; }
 
-        Environment GetEnvironmentById(int id);
+        Environment GetEnvironmentById(long id);
